move statusClass to helper
